+++ test/filtrations/new-filtrations.jl
@@ -75,13 +75,13 @@ end
             appa = ApparentPairsRips(projective_plane)
             _, hom_imp = ripserer(appa; alg=:homology, implicit=true, modulus=m)
             _, hom_exp = ripserer(appa; alg=:homology, implicit=false, modulus=m)
-            _, coh_imp = ripserer(appa; alg=:cohomology, implicit=true, modulus=m)
-            _, coh_exp = ripserer(appa; alg=:cohomology, implicit=false, modulus=m)
+            _, hom_ass = ripserer(appa; alg=:involuted, implicit=false, modulus=m)
+            _, coh_imp = ripserer(appa; alg=:cohomology, implicit=true, modulus=m, reps=1)
+            _, coh_exp = ripserer(appa; alg=:cohomology, implicit=false, modulus=m, reps=1)
 
-            @test hom_imp == hom_exp == coh_imp == coh_exp
-            @test all(
-                i.representative == j.representative for (i, j) in zip(hom_imp, hom_exp)
-            )
+            @test hom_imp == hom_exp == coh_imp == coh_exp == hom_ass
+            @test representative.(hom_imp) == representative.(hom_exp)
+            @test representative.(hom_imp) == representative.(hom_ass)
         end
     end
 end
@@ -104,7 +104,7 @@ Ripserer.adjacency_matrix(::ApparentPairsCustom) = sparse(Bool[0 1 1; 1 0 1; 1 1
 function Ripserer.find_apparent_pairs(a::ApparentPairsCustom, columns, _)
     σ = columns[1]
     a.found[] = true
-    return typeof(σ)[], [(σ, minimum(Ripserer.coboundary(a, σ)))]
+    return typeof(σ)[], [(σ, only(Ripserer.coboundary(a, σ)))]
 end
 
 @testset "Apparent pairs can produce valid intervals." begin
@@ -127,7 +127,7 @@ end
         _, coh_exp = ripserer(appa; alg=:cohomology, implicit=false)
 
         @test hom_imp == hom_exp == coh_imp == coh_exp
-        @test all(i.representative == j.representative for (i, j) in zip(hom_imp, hom_exp))
+        @test representative.(hom_imp) == representative.(hom_exp)
     end
 end